make Dom a shared component
